Edit Site: Move styles preview colors to a hook
